Improve the overall design for shoesrb website.
The homepage now contains 4 major pages:

* The landing/home page.
* Contributers page for people interested in contributing.
* Tutorials page for people interested in learning Shoes.
* Blog

The About page is empty for now but would be great to have some information
about Shoes here.

Additional Note: The texts for the page are not that great. Partly because
I had English as a second language in school :) So improvements are welcome.