Add downloads
Add a downloads page and a handler for individual downloads.
The download handler translates a URL that looks like this:

    http://shoesrb.com/downloads/filename.zip

into a redirect to an external URL, where the file is actually located.
To add a new download, just map the local filename to the remote url
in download_resource.rb.